Fixed registration and login error handling. Improved tests to do error checking.
