+++ server/Routes/Auth/admin.py
@@ -16,4 +16,4 @@ class AdminRoute(Resource):
             result = adminmanager.login(request.json['password'])
         if result:
             return make_response(jsonify({'token': result}), HTTPStatus.OK)
-        abort(HTTPStatus.NOT_FOUND, 'User does not exist or password is wrong')
+        abort(HTTPStatus.UNPROCESSABLE_ENTITY, 'User does not exist or password is wrong')
