+++ server/Routes/Auth/register.py
@@ -1,5 +1,5 @@
 from http import HTTPStatus
-from flask import make_response, request, abort
+from flask import make_response, request, abort, jsonify
 from flask_restful import Resource
 
 from server.Managers.UserManager import UserManager
@@ -12,5 +12,6 @@ class RegisterRoute(Resource):
         with UserManager(request.json['email']) as usermanager:
             result = usermanager.register(request.json['password'])
         if result:
-            return make_response('Account partially created. Verification email sent\n', HTTPStatus.OK)
+            return make_response(jsonify({'message': 'Account partially created. Verification email sent\n'}), 
+                                HTTPStatus.OK)
         abort(HTTPStatus.UNPROCESSABLE_ENTITY, 'User already exists')
