+++ tests/ref/test_general_login_non_existing.ref
@@ -1,0 +1,4 @@
+{
+    "message": "Username or Password is wrong"
+}
+HTTP_Status: 422
